Revive project--create server
This commit causes LizardServer.py to implement an actual server.
It listens on port 13001 for connections. If the message received
from the client begins with 'REGISTER', the next argument is
interpreted as a firebase device id token and the device is given
a unique device id name (deviceN, where N is an integer). If the
message received begins with 'SEND_MESSAGE', the next argument is
interpreted as a device name (deviceN), and the following is
interpreted as the message. The Lizard app has been configured to
register devices with the server edtwardy.hopto.org. The script
SendMessage.py can be used to send push notifications to devices.